Add glass pyramid water flow functionality

+++ src/water_overflow/liter_to_ml.py
@@ -1,7 +1,7 @@
 from src.errors.liters_conversion_error import LitersConversionError
 
 
-def liter_to_ml(liter: int) -> int:
+def liter_to_ml(liter: float) -> float:
     if liter < 0:
         raise LitersConversionError()
     return liter * 1000
